Adding Logic to Drug Category Service
